Replace all float with double

+++ archived/window.cc
@@ -1,6 +1,6 @@
 #include "window.h"
 
-void Window::applyWindow(float* input) {
+void Window::applyWindow(double* input) {
     for (int i=0; i<length; ++i)
         input[i]*=_window[i];
 }
